+++ libsrc/ffdec_lib/src/com/jpexs/decompiler/flash/exporters/shape/BitmapExporter.java
@@ -249,7 +249,7 @@ public class BitmapExporter extends ShapeExporterBase {
         finalizePath();
         if (color == null) {
             fillPaint = defaultColor;
-        } else {
+        } else {            
             fillPaint = color.toColor();
         }
     }